+++ AddTeacher.cs
@@ -17,7 +17,6 @@ namespace CMS
         public AddTeacher()
         {
             InitializeComponent();
-            sqlConnection.Open();
         }
 
         void ShowData()
@@ -27,14 +26,29 @@ namespace CMS
             sqlDataAdapter.Fill(dataTable);
 
             TeacherTable.DataSource = dataTable;
+
+            SqlCommand sqlCommand = new SqlCommand("SELECT department FROM DEPARTMENT", sqlConnection);
+            SqlDataReader sqlDataReader = sqlCommand.ExecuteReader();
+            while (sqlDataReader.Read())
+            {
+                ComboDepartment.Items.Add(sqlDataReader.GetString(0));
+            }
+            sqlDataReader.Close();
         }
 
         private void AddTeacher_Load(object sender, EventArgs e)
         {
+
             BtnVisibility();
             try
             {
-                ShowData();
+                if (sqlConnection.State == ConnectionState.Closed)
+                {
+                    sqlConnection.Open();
+                    ShowData();
+                    sqlConnection.Close();
+                }
+
             }
             catch(Exception ex) {
                 MessageBox.Show(ex.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
@@ -60,24 +74,32 @@ namespace CMS
 
         private void BtnSave_Click(object sender, EventArgs e)
         {
-            try {
+
+            try
+            {
                 if (TxtTeacherName.TextLength > 0 && TxtMobNumber.TextLength > 0 && TxtAddress.TextLength > 0 && TxtEmail.TextLength > 0)
                 {
-                    string query = "INSERT INTO TEACHER(tname, tgender, tdob, tphone, taddress, temail, tdepartment)VALUES(@tname,@tgender,@tdob,@tphone,@taddress,@temail,@tdepartment)";
-                    SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
-                    sqlCommand.Parameters.AddWithValue("@tname", TxtTeacherName.Text);
-                    sqlCommand.Parameters.AddWithValue("@tgender", ComboGender.Text);
-                    sqlCommand.Parameters.AddWithValue("@tdob", PickerDateTime.Value.Date);
-                    sqlCommand.Parameters.AddWithValue("@tphone", TxtMobNumber.Text);
-                    sqlCommand.Parameters.AddWithValue("@taddress", TxtAddress.Text);
-                    sqlCommand.Parameters.AddWithValue("@temail", TxtEmail.Text);
-                    sqlCommand.Parameters.AddWithValue("@tdepartment", ComboDepartment.Text);
-                    sqlCommand.ExecuteNonQuery();
+                    if (sqlConnection.State != ConnectionState.Open)
+                    {
+                        sqlConnection.Open();
+                        string query = "INSERT INTO TEACHER(tname, tgender, tdob, tphone, taddress, temail, tdepartment)VALUES(@tname,@tgender,@tdob,@tphone,@taddress,@temail,@tdepartment)";
+                        SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@tname", TxtTeacherName.Text);
+                        sqlCommand.Parameters.AddWithValue("@tgender", ComboGender.Text);
+                        sqlCommand.Parameters.AddWithValue("@tdob", PickerDateTime.Value.Date);
+                        sqlCommand.Parameters.AddWithValue("@tphone", TxtMobNumber.Text);
+                        sqlCommand.Parameters.AddWithValue("@taddress", TxtAddress.Text);
+                        sqlCommand.Parameters.AddWithValue("@temail", TxtEmail.Text);
+                        sqlCommand.Parameters.AddWithValue("@tdepartment", ComboDepartment.Text);
+                        sqlCommand.ExecuteNonQuery();
 
-                    ShowData();
+                        ShowData();
 
-                    MessageBox.Show("Teacher Added Successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    Reset();                
+                        MessageBox.Show("Teacher Added Successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        Reset();
+                        sqlConnection.Close();
+                    }
+                                   
                 }
                 else
                 {
@@ -90,6 +112,7 @@ namespace CMS
 
             }
             BtnVisibility();
+
         }
 
         void Reset()
@@ -122,26 +145,32 @@ namespace CMS
 
         private void BtnUpdate_Click(object sender, EventArgs e)
         {
+
             try
             {
                 if (TxtId.TextLength > 0 && TxtTeacherName.TextLength > 0 && TxtMobNumber.TextLength > 0 && TxtAddress.TextLength > 0 && TxtEmail.TextLength > 0)
                 {
-                    string query = "UPDATE TEACHER SET tname=@name, tgender=@gender, tdob=@dob, tphone=@phone, temail=@email, tdepartment=@department, taddress=@address WHERE tid = @id";
-                    SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
-                    sqlCommand.Parameters.AddWithValue("@id", TxtId.Text);  
-                    sqlCommand.Parameters.AddWithValue("@name", TxtTeacherName.Text);
-                    sqlCommand.Parameters.AddWithValue("@gender", ComboGender.Text);
-                    sqlCommand.Parameters.AddWithValue("@dob", PickerDateTime.Value.Date);
-                    sqlCommand.Parameters.AddWithValue("@phone", TxtMobNumber.Text);
-                    sqlCommand.Parameters.AddWithValue("@email", TxtEmail.Text);
-                    sqlCommand.Parameters.AddWithValue("@department", ComboDepartment.Text);
-                    sqlCommand.Parameters.AddWithValue("@address", TxtAddress.Text);
-                    sqlCommand.ExecuteNonQuery();
+                    if (sqlConnection.State == ConnectionState.Closed)
+                    {
+                        sqlConnection.Open();
+                        string query = "UPDATE TEACHER SET tname=@name, tgender=@gender, tdob=@dob, tphone=@phone, temail=@email, tdepartment=@department, taddress=@address WHERE tid = @id";
+                        SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@id", TxtId.Text);
+                        sqlCommand.Parameters.AddWithValue("@name", TxtTeacherName.Text);
+                        sqlCommand.Parameters.AddWithValue("@gender", ComboGender.Text);
+                        sqlCommand.Parameters.AddWithValue("@dob", PickerDateTime.Value.Date);
+                        sqlCommand.Parameters.AddWithValue("@phone", TxtMobNumber.Text);
+                        sqlCommand.Parameters.AddWithValue("@email", TxtEmail.Text);
+                        sqlCommand.Parameters.AddWithValue("@department", ComboDepartment.Text);
+                        sqlCommand.Parameters.AddWithValue("@address", TxtAddress.Text);
+                        sqlCommand.ExecuteNonQuery();
 
-                    ShowData();
+                        ShowData();
 
-                    MessageBox.Show("Teacher Data Updated Successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    Reset();
+                        MessageBox.Show("Teacher Data Updated Successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        Reset();
+                        sqlConnection.Close();
+                    }
                 }
                 else
                 {
@@ -162,19 +191,25 @@ namespace CMS
 
         private void BtnDelete_Click(object sender, EventArgs e)
         {
+
             try
             {
                 if (TxtId.TextLength > 0 && TxtTeacherName.TextLength > 0 && TxtMobNumber.TextLength > 0 && TxtAddress.TextLength > 0 && TxtEmail.TextLength > 0)
                 {
-                    string query = "DELETE FROM TEACHER WHERE tid=@id";
-                    SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
-                    sqlCommand.Parameters.AddWithValue("@id", TxtId.Text);
-                    sqlCommand.ExecuteNonQuery();
+                    if (sqlConnection.State == ConnectionState.Closed)
+                    {
+                        sqlConnection.Open();
+                        string query = "DELETE FROM TEACHER WHERE tid=@id";
+                        SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@id", TxtId.Text);
+                        sqlCommand.ExecuteNonQuery();
 
-                    ShowData();
+                        ShowData();
 
-                    MessageBox.Show("Teacher Data Deleted Successfully.", "Updated", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    Reset();
+                        MessageBox.Show("Teacher Data Deleted Successfully.", "Updated", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        Reset();
+                        sqlConnection.Close();
+                    }
                 }
                 else
                 {
@@ -188,6 +223,7 @@ namespace CMS
             {
                 MessageBox.Show(ex.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
+
         }
     }
 }
